# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Jobs",
        "description": "Laravel API documentation for the 'My Jobs' application.",
        "contact": {
            "email": "erick1souza1ago04@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/vacancies/{id}": {
            "get": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Retrieve a single vacancy",
                "description": "Retrieve details of a single vacancy by its ID.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vacancy to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vacancy details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID of the vacancy.",
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "short_description": {
                                                    "description": "Short description of the vacancy.",
                                                    "type": "string",
                                                    "example": "Test"
                                                },
                                                "long_description": {
                                                    "description": "Long description of the vacancy.",
                                                    "type": "string",
                                                    "example": "Test long description"
                                                },
                                                "wage": {
                                                    "description": "Wage of the vacancy.",
                                                    "type": "string",
                                                    "example": "R$ 12409,14"
                                                },
                                                "zip_code": {
                                                    "description": "Zip code (CEP) of the vacancy.",
                                                    "type": "string",
                                                    "example": "14148-300"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "description": "ID of the user.",
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "name": {
                                                            "description": "Name of the user.",
                                                            "type": "string",
                                                            "example": "Erick"
                                                        },
                                                        "email": {
                                                            "description": "Email of the user.",
                                                            "type": "string",
                                                            "example": "erick@jobs.com"
                                                        },
                                                        "email_verified_at": {
                                                            "description": "Email verified at of the user.",
                                                            "type": "string",
                                                            "example": "erick_verify@jobs.com"
                                                        },
                                                        "created_at": {
                                                            "description": "created at of the user.",
                                                            "type": "string",
                                                            "example": "2024-06-04T01:42:02.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "description": "updated at of the user.",
                                                            "type": "string",
                                                            "example": "2024-06-04T01:42:02.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "zip": {
                                                    "properties": {
                                                        "zip_code": {
                                                            "description": "code of the zip.",
                                                            "type": "string",
                                                            "example": "14085-440"
                                                        },
                                                        "street": {
                                                            "description": "street of the zip.",
                                                            "type": "string",
                                                            "example": "Captain Solomon Avenue"
                                                        },
                                                        "complement": {
                                                            "description": "Complement of the zip.",
                                                            "type": "string",
                                                            "example": "Close to the market x"
                                                        },
                                                        "neighborhood": {
                                                            "description": "neighborhood of the zip.",
                                                            "type": "string",
                                                            "example": "Elysian Fields"
                                                        },
                                                        "locality": {
                                                            "description": "city of the zip.",
                                                            "type": "string",
                                                            "example": "Central City"
                                                        },
                                                        "uf": {
                                                            "description": "uf the zip.",
                                                            "type": "string",
                                                            "example": "SP"
                                                        },
                                                        "ibge": {
                                                            "description": "ibge code of the zip.",
                                                            "type": "integer",
                                                            "example": "3543402"
                                                        },
                                                        "gia": {
                                                            "description": "gia code of the zip.",
                                                            "type": "integer",
                                                            "example": "5824"
                                                        },
                                                        "ddd": {
                                                            "description": "ddd code of the zip.",
                                                            "type": "integer",
                                                            "example": "16"
                                                        },
                                                        "siafi": {
                                                            "description": "siafi code of the zip.",
                                                            "type": "integer",
                                                            "example": "6969"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vacancy not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Update an existing vacancy",
                "description": "Update an existing vacancy with the provided data.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vacancy to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "short_description": {
                                        "description": "Short description of the vacancy (between 5 and 60 characters).",
                                        "type": "string",
                                        "example": "Updated short description"
                                    },
                                    "long_description": {
                                        "description": "Long description of the vacancy (between 10 and 250 characters).",
                                        "type": "string",
                                        "example": "Updated long description"
                                    },
                                    "wage": {
                                        "description": "Wage of the vacancy (in cents).",
                                        "type": "integer",
                                        "example": 1500000
                                    },
                                    "zip_code": {
                                        "description": "Zip code (CEP) of the vacancy.",
                                        "type": "string",
                                        "example": "11020-123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacancy updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID of the updated vacancy.",
                                                    "type": "integer",
                                                    "example": "8"
                                                },
                                                "short_description": {
                                                    "description": "Short description of the vacancy.",
                                                    "type": "string",
                                                    "example": "Updated short description"
                                                },
                                                "long_description": {
                                                    "description": "Long description of the vacancy.",
                                                    "type": "string",
                                                    "example": "Updated long description"
                                                },
                                                "wage": {
                                                    "description": "Wage of the vacancy.",
                                                    "type": "string",
                                                    "example": "1500000"
                                                },
                                                "zip_code": {
                                                    "description": "Zip code (CEP) of the vacancy.",
                                                    "type": "string",
                                                    "example": "11020-123"
                                                },
                                                "user_id": {
                                                    "description": "ID of the user updating the vacancy.",
                                                    "type": "integer",
                                                    "example": "5"
                                                },
                                                "updated_at": {
                                                    "description": "Date and time of the last update.",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-05-17T12:00:00.000000Z"
                                                },
                                                "created_at": {
                                                    "description": "Date and time of creation.",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-05-17T00:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - The user is not authorized to update this vacancy."
                    },
                    "404": {
                        "description": "Vacancy not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Delete a vacancy",
                "description": "Delete a vacancy by its ID.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vacancy to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vacancy deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Confirmation message.",
                                            "type": "string",
                                            "example": "Vacancy successfully deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - The user is not authorized to delete this vacancy."
                    },
                    "404": {
                        "description": "Vacancy not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}